fix: wrong prototype

+++ src/main.cpp
@@ -10,7 +10,7 @@
 
 string get(string name,string default_parameter);
 double get_double(string name,string default_parameter);
-double get_int(string name,string default_parameter);
+int get_int(string name,string default_parameter);
 void help_message();
 bool restoration();
